* task-013: (35 commits)
  Change address literals from 16 to 12 bits on tests
  Fix idiom while decoding 'ldVxK' AGAIN
  Removed a zero from 'i' register initialization
  Fix idiom while decoding 'ldVxK'
  Fix compilation of the 'console' project
  Modify the 'ldVxK' instruction to play sound
  Fix Keyboad initialization in Chipeit class
  Add the 'ICoreKeyboard' interface
  Fix typo in 'LoggedKeybard' class
  Rename IKeyboard to IUserKeyboard
  Clarify that PC and SP are addresses
  Convert single-multi-line comment to multiline
  Apply conventions for registers manipulation
  Adapt instruction decoding to secret convention
  Remove wrapping 'when' expression inside a lambda
  Guarantee a stack of a minimum size of 16
  Replace 'thenAnswer' for 'thenReturn' in tests
  Add missing 'ldVxK' invocation from the Cpu
  Unified V registers indexes criteria
  Unified instructions and tests style
  ...

+++ src/chipeitlib/src/main/java/es/chipeit/lib/interfaces/IRegisters.kt
@@ -5,5 +5,5 @@ internal interface IRegisters {
     var i: Short
     var pc: Int
 
-    var sp: Byte
+    var sp: Int
 }
